+++ src/main/java/com/lizekai/wms/service/impl/RoleServiceImpl.java
@@ -7,6 +7,7 @@ import com.lizekai.wms.constants.SystemCanstants;
 import com.lizekai.wms.domain.ResponseResult;
 import com.lizekai.wms.domain.entity.RoleMenu;
 import com.lizekai.wms.domain.vo.PageVo;
+import com.lizekai.wms.handler.exception.SystemException;
 import com.lizekai.wms.mapper.RoleMapper;
 import com.lizekai.wms.domain.entity.Role;
 import com.lizekai.wms.service.RoleMenuService;
@@ -80,6 +81,10 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements Ro
 
     @Override
     public ResponseResult updateRole(Role role) {
+        //不能修改超级管理员
+        if(SystemCanstants.IS_ADMIN.equals(role.getId().toString())){
+            return ResponseResult.errorResult(500,"不能修改超级管理员");
+        }
         updateById(role);
         roleMenuService.deleteRoleMenuByRoleId(role.getId());
         insertRoleMenu(role);
